resolve namespace and pickling errors
