+++ src/spyglass/utils/dj_helper_fn.py
@@ -466,3 +466,37 @@ def make_file_obj_id_unique(nwb_path: str):
         f.attrs["object_id"] = new_id
     _resolve_external_table(nwb_path, nwb_path.split("/")[-1])
     return new_id
+
+def populate_pass_function(value):
+    """Pass function for parallel populate.
+
+    Note: To avoid pickling errors, the table must be passed by class, NOT by instance.
+    Note: This function must be defined in the global namespace.
+
+    Parameters
+    ----------
+    value : (table, key, kwargs)
+       Class of table to populate, key to populate, and kwargs fro populate
+    """
+    table, key, kwargs = value
+    return table.populate(key, **kwargs)
+
+class NonDaemonPool(multiprocessing.pool.Pool):
+    """NonDaemonPool. Used to create a pool of non-daemonized processes,
+    which are required for parallel populate operations in DataJoint.
+    """
+    def Process(self, *args, **kwds):
+        proc = super(NonDaemonPool, self).Process(*args, **kwds)
+
+        class NonDaemonProcess(proc.__class__):
+            """Monkey-patch process to ensure it is never daemonized"""
+            @property
+            def daemon(self):
+                return False
+
+            @daemon.setter
+            def daemon(self, val):
+                pass
+
+        proc.__class__ = NonDaemonProcess
+        return proc
