+++ src/spyglass/utils/dj_mixin.py
@@ -20,7 +20,7 @@ from networkx import NetworkXError
 from pymysql.err import DataError
 
 from spyglass.utils.database_settings import SHARED_MODULES
-from spyglass.utils.dj_helper_fn import fetch_nwb, get_nwb_table
+from spyglass.utils.dj_helper_fn import fetch_nwb, get_nwb_table, populate_pass_function, NonDaemonPool
 from spyglass.utils.dj_merge_tables import RESERVED_PRIMARY_KEY as MERGE_PK
 from spyglass.utils.dj_merge_tables import Merge, is_merge_table
 from spyglass.utils.logging import logger
@@ -663,6 +663,11 @@ class SpyglassMixin:
 
     # -------------------------- non-daemon populate --------------------------
     def populate(self, *restrictions, **kwargs):
+        """Populate table in parallel.
+
+        Superceeds datajoint.table.Table.populate for classes with that
+        spawn processes in their make function
+        """
 
         # Pass through to super if not parallel in the make function or only a single process
         processes = kwargs.pop("processes", 1)
@@ -670,38 +675,21 @@ class SpyglassMixin:
             return super().populate(*restrictions, **kwargs)
 
         # If parallel in both make and populate, use non-daemon processes
+        # Get keys to populate
         keys = (self._jobs_to_do(restrictions) - self.target).fetch(
             "KEY", limit=kwargs.get("limit", None)
         )
+        # package the call list
+        call_list = [(type(self), key, kwargs) for key in keys]
 
-        class NonDaemonPool(multiprocessing.pool.Pool):
-            def Process(self, *args, **kwds):
-                proc = super(NonDaemonPool, self).Process(*args, **kwds)
-
-                class NonDaemonProcess(proc.__class__):
-                    print("HI")
-                    """Monkey-patch process to ensure it is never daemonized"""
-
-                    @property
-                    def daemon(self):
-                        return False
-
-                    @daemon.setter
-                    def daemon(self, val):
-                        pass
-
-                proc.__class__ = NonDaemonProcess
-                return proc
-
-        pass_function = lambda key: super().populate(key, **kwargs)
+        # Create a pool of non-daemon processes to populate a single entry each
         pool = NonDaemonPool(processes=processes)
         try:
-            pool.map(pass_function, keys)
+            pool.map(populate_pass_function, call_list)
         except Exception as e:
             pool.close()
             pool.terminate()
             raise e
-
         pool.close()
         pool.terminate()
